Feature: Allow copying the table definition from the dropdown menu
